04_auditez_un_environnement_de_donnees/pyproject.toml
[tool.poetry]
name = "04-auditez-un-environnement-de-donnees"
version = "0.1.0"
description = ""
authors = ["joaquim.diaz <joaquim.diaz@orange.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
polars = "^1.11.0"
fastexcel = "^0.12.0"
ipykernel = "^6.29.5"
unidecode = "^1.3.8"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
cryptography = "^44.0.0"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
